So, I went on some massive crusade to fix hot reloading
I thought something was wrong with the boiler, so I replaced it with
https://github.com/facebookincubator/create-react-app . Same problem. It
turns out that vim and and way that webpack does file change detection
is incompatible, and has been since 2015. It's fixed in Webpack 2.

There's a Vim workaround, but now I'm trying on Atom + vim mode for size.